Refactor manfifest repo dependency injection to common function
