+++ tests/test_manifests.py
@@ -2,23 +2,15 @@ import pytest
 from fastapi.testclient import TestClient
 
 from alwayson.app import app
-from alwayson.controllers.manifests import get_manifest_repo
-from alwayson.db.manifest_repo import ManifestRepo
+from alwayson.controllers.injectors import inject_manifest_repo
 from alwayson.models import AlwaysOnRequest, AlwaysOnResponse, ManifestEntry
-
-
-@pytest.fixture
-def mock_manifest_repo(mock_db):
-  def create_mock():
-    return ManifestRepo(mock_db)
-  return create_mock
 
 
 @pytest.mark.asyncio
 async def test_upload_full_manifest(mock_db, mock_manifest_repo):
   # arrange
   subject = TestClient(app)
-  app.dependency_overrides[get_manifest_repo] = mock_manifest_repo
+  app.dependency_overrides[inject_manifest_repo] = mock_manifest_repo
 
   payload = [
     {
